requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
greenlet==3.0.1
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
parse==1.20.0
parse-type==0.6.2
playwright==1.40.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pyee==11.0.1
pytest==7.4.3
pytest-base-url==2.0.0
pytest-bdd==7.0.0
pytest-playwright==0.4.3
python-slugify==8.0.1
requests==2.31.0
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
